Refactored the worldpop-count sub-pipeline into a module
