+++ src/dart_pipeline/INDEX.md
@@ -1,6 +1,20 @@
 INDEX
 =====
 
+`geospatial`
+
+```
+worldpop_count.py
+✅ def process_gadm_worldpopcount(
+```
+
+`sociodemographic`
+
+```
+worldpop_count.py
+✅ def process_gadm_worldpopcount(
+```
+
 `__init__.py`
 
 ```
@@ -41,14 +55,12 @@ def get_admin(x):
 ✅ def process_dengueperu(
 ✅ def process_gadm_aphroditetemperature(
 def process_gadm_admin_map_data(iso3: str, admin_level: AdminLevel):
-def process_gadm_worldpopcount(
 ✅ def process_aphrodite_temperature_data(year=None, plots=False) -> \
 ✅ def process_aphrodite_precipitation_data(
 def get_chirps_rainfall_data_path(date: PartialDate) -> Path:
 def process_chirps_rainfall(partial_date: str, plots=False) -> ProcessResult:
 def process_era5_reanalysis_data() -> ProcessResult:
 def process_terraclimate(
-def process_worldpop_pop_count_data(
 def process_worldpop_pop_density_data(iso3: str, year: int) -> ProcessResult:
 def process_gadm_chirps_rainfall(
 def get_admin_region(lat: float, lon: float, polygons) -> str:
