+++ src/dart_pipeline/constants.py
@@ -15,6 +15,9 @@ COL_OFF = "\033[0m"
 MSG_SOURCE = COL_BLUE + "  source" + COL_OFF
 MSG_PROCESS = COL_CYAN + " process" + COL_OFF
 
+# Smallest single-precision floating-point number
+MIN_FLOAT = -3.4028234663852886e38
+
 TERRACLIMATE_METRICS = [
     "aet",  # water_evaporation_amount_mm
     "def",  # water_potential_evaporation_amount_minus_water_evaporatio
